Add test typecheck gate and split oversized test contracts (#83)
Introduce npm run typecheck:tests over all test/**/*.ts, tighten fixtures
with SDK/pi typing helpers, and split the largest provider/index/discovery/
transcript suites by public contract without changing runtime behavior.
