+++ scripts/lib/cursor-child-process.d.mts
@@ -1,0 +1,10 @@
+import type { ChildProcess } from "node:child_process";
+
+export declare const DEFAULT_CHILD_SHUTDOWN_GRACE_MS: number;
+export declare function waitForChildClose(child: ChildProcess): Promise<void>;
+export declare function signalChild(child: ChildProcess, signal: NodeJS.Signals): void;
+export declare function terminateChild(
+	child: ChildProcess,
+	options?: { graceMs?: number },
+): Promise<void>;
+export declare function parseJsonLines(stdout: string): unknown[];
