+++ scripts/lib/cursor-script-fail.d.mts
@@ -1,0 +1,1 @@
+export declare function createScriptFail(prefix: string): (message: string, secrets?: string | string[]) => never;
