+++ test/cursor-provider-stream-incomplete-tools.test.ts
@@ -1,0 +1,293 @@
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import { Type } from "typebox";
+import {
+	resetCursorProviderTestState,
+	mockedCreate,
+	mockedCreateAgentPlatform,
+	makeModel,
+	makeContext,
+	makeAssistantMessage,
+	collectEvents,
+	collectTextDeltas,
+	collectThinkingDeltas,
+	getEventsOfType,
+	getDoneEvent,
+	getErrorEvent,
+	getTextEndEvent,
+	hasEventType,
+	isToolCallBlock,
+	isCursorToolStreamEvent,
+	getCreatedAgentOptions,
+	createMockAgentPlatform,
+	registerBridgeForProviderTest,
+	registerNativeToolDisplayForTest,
+	connectMcpClient,
+	createBuiltinToolInfo,
+	createTestToolInfo,
+	cursorModelItems,
+	type CursorDeltaHandler,
+	type CursorStepHandler,
+	type RegisteredTool,
+	mockCreatedAgent,
+	asMockCursorRun,
+	getPiToolsMcpUrlFromAgentCreateOptions,
+} from "./helpers/cursor-provider-harness.js";
+import { streamCursor, __testUtils as cursorProviderTestUtils } from "../src/cursor-provider.js";
+import { __testUtils as contextWindowCacheTestUtils } from "../src/context-window-cache.js";
+import { __testUtils as modelDiscoveryTestUtils } from "../src/model-discovery.js";
+import { __testUtils as sdkEventDebugTestUtils } from "../src/cursor-sdk-event-debug.js";
+import type { SDKMessage, SendOptions } from "@cursor/sdk";
+import type { Context } from "@earendil-works/pi-ai";
+
+type CursorOnStepPayload = Parameters<NonNullable<SendOptions["onStep"]>>[0];
+import { mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+
+
+
+
+describe("streamCursor incomplete tools", () => {
+	beforeEach(resetCursorProviderTestState);
+
+		it("surfaces incomplete started Cursor tool calls with neutral activity traces", async () => {
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({ update: { type: "tool-call-started", toolCall: { name: "shell", args: { command: "sleep 10" } }, callId: "c1" } });
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			const stream = streamCursor(makeModel(), makeContext(), { apiKey: "test-key" });
+			const events = await collectEvents(stream);
+			const trace = collectThinkingDeltas(events);
+			const text = collectTextDeltas(events);
+	
+			expect(trace).toContain("Cursor shell did not complete");
+			expect(trace).toContain("missing completion");
+			expect(text).toBe("done");
+			expect(hasEventType(events, "toolcall_start")).toBe(false);
+		});
+	
+		it("surfaces incomplete Cursor web search MCP activity with a distinct label", async () => {
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({
+					update: {
+						type: "tool-call-started",
+						toolCall: { name: "mcp", args: { toolName: "WebSearch", args: { search_term: "pi extension" } } },
+						callId: "c1",
+					},
+				});
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			const events = await collectEvents(streamCursor(makeModel(), makeContext(), { apiKey: "test-key" }));
+			const trace = collectThinkingDeltas(events);
+			expect(trace).toContain("Cursor web search did not complete");
+			expect(trace).not.toContain("Cursor MCP did not complete");
+		});
+	
+		it("surfaces incomplete generic Cursor MCP activity", async () => {
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({
+					update: {
+						type: "tool-call-started",
+						toolCall: { name: "mcp", args: { toolName: "git" } },
+						callId: "c1",
+					},
+				});
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			const events = await collectEvents(streamCursor(makeModel(), makeContext(), { apiKey: "test-key" }));
+			expect(collectThinkingDeltas(events)).toContain("Cursor MCP did not complete");
+		});
+	
+		it("records discarded incomplete started tool calls to coordinator-events.jsonl when PI_CURSOR_SDK_EVENT_DEBUG is enabled", async () => {
+			const artifactDir = mkdtempSync(join(tmpdir(), "pi-cursor-sdk-provider-discarded-debug-"));
+			process.env.PI_CURSOR_SDK_EVENT_DEBUG = "1";
+			process.env.PI_CURSOR_SDK_EVENT_DEBUG_RUN_DIR = artifactDir;
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({ update: { type: "tool-call-started", toolCall: { name: "read", args: { path: "README.md" } }, callId: "c1" } });
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			try {
+				await collectEvents(streamCursor(makeModel(), makeContext(), { apiKey: "test-key" }));
+				const coordinatorEvents = readFileSync(join(artifactDir, "coordinator-events.jsonl"), "utf8");
+				expect(coordinatorEvents).toContain("discarded-incomplete-started-tool-call");
+				expect(coordinatorEvents).toContain('"toolName":"read"');
+				expect(coordinatorEvents).toContain('"reason":"no-completion-at-run-end"');
+				expect(coordinatorEvents).not.toContain("c1");
+			} finally {
+				delete process.env.PI_CURSOR_SDK_EVENT_DEBUG;
+				delete process.env.PI_CURSOR_SDK_EVENT_DEBUG_RUN_DIR;
+				rmSync(artifactDir, { recursive: true, force: true });
+			}
+		});
+	
+		it("suppresses incomplete missing-file reads with final error text while keeping debug evidence", async () => {
+			const artifactDir = mkdtempSync(join(tmpdir(), "pi-cursor-sdk-provider-missing-read-debug-"));
+			process.env.PI_CURSOR_SDK_EVENT_DEBUG = "1";
+			process.env.PI_CURSOR_SDK_EVENT_DEBUG_RUN_DIR = artifactDir;
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({ update: { type: "tool-call-started", toolCall: { name: "read", args: { path: "missing.txt" } }, callId: "c-missing" } });
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "Error: File not found" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			try {
+				const events = await collectEvents(streamCursor(makeModel(), makeContext(), { apiKey: "test-key" }));
+				const trace = collectThinkingDeltas(events);
+				const text = collectTextDeltas(events);
+				const coordinatorEvents = readFileSync(join(artifactDir, "coordinator-events.jsonl"), "utf8");
+				const displayDecisions = readFileSync(join(artifactDir, "display-decisions.jsonl"), "utf8");
+	
+				expect(text).toBe("Error: File not found");
+				expect(trace).not.toContain("Cursor read did not complete");
+				expect(hasEventType(events, "toolcall_start")).toBe(false);
+				expect(coordinatorEvents).toContain("discarded-incomplete-started-tool-call");
+				expect(coordinatorEvents).toContain('"toolName":"read"');
+				expect(coordinatorEvents).not.toContain("c-missing");
+				expect(displayDecisions).toContain('"action":"skip-incomplete-fast-local"');
+				expect(displayDecisions).toContain('"toolName":"read"');
+			} finally {
+				delete process.env.PI_CURSOR_SDK_EVENT_DEBUG;
+				delete process.env.PI_CURSOR_SDK_EVENT_DEBUG_RUN_DIR;
+				rmSync(artifactDir, { recursive: true, force: true });
+			}
+		});
+	
+		it("still surfaces explicit completed Cursor tool errors", async () => {
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler }) => {
+				opts.onDelta({ update: { type: "tool-call-started", toolCall: { name: "shell", args: { command: "cat missing.txt" } }, callId: "c1" } });
+				opts.onDelta({
+					update: {
+						type: "tool-call-completed",
+						toolCall: {
+							name: "shell",
+							args: { command: "cat missing.txt" },
+							result: { status: "error", error: "missing.txt: No such file" },
+						},
+						callId: "c1",
+					},
+				});
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			const stream = streamCursor(makeModel(), makeContext(), { apiKey: "test-key" });
+			const events = await collectEvents(stream);
+			const trace = collectThinkingDeltas(events);
+	
+			expect(trace).toContain("$ cat missing.txt");
+			expect(trace).toContain("Error: missing.txt: No such file");
+		});
+	
+		it("still surfaces explicit onStep Cursor tool errors", async () => {
+			const mockSend = vi.fn().mockImplementation(async (_msg: unknown, opts: { onDelta: CursorDeltaHandler; onStep: CursorStepHandler }) => {
+				opts.onDelta({ update: { type: "tool-call-started", toolCall: { name: "read", args: { path: "missing.txt" } }, callId: "c1" } });
+				opts.onStep({
+					step: {
+						type: "toolCall",
+						id: "c1",
+						message: {
+							type: "read",
+							args: { path: "missing.txt" },
+							result: { status: "error", error: "missing.txt: No such file" },
+						},
+					} as CursorOnStepPayload["step"],
+				});
+				return asMockCursorRun({
+					id: "run-1",
+					agentId: "agent-1",
+					status: "finished",
+					wait: vi.fn().mockResolvedValue({ id: "run-1", status: "finished", result: "done" }),
+					cancel: vi.fn(),
+					supports: () => true,
+					unsupportedReason: () => undefined,
+				});
+			});
+			mockCreatedAgent({
+				send: mockSend,
+				[Symbol.asyncDispose]: vi.fn().mockResolvedValue(undefined),
+			});
+	
+			const stream = streamCursor(makeModel(), makeContext(), { apiKey: "test-key" });
+			const events = await collectEvents(stream);
+			const trace = collectThinkingDeltas(events);
+	
+			expect(trace).toContain("read missing.txt");
+			expect(trace).toContain("Error: missing.txt: No such file");
+			expect(trace).not.toContain("Cursor tool started without a completion event");
+		});
+	
+});
